+++ sbt-scalajs-bundler/src/sbt-test/sbt-scalajs-bundler/library/src/main/scala/example/Library.scala
@@ -1,9 +1,9 @@
 //#library-definition
 package example
 
-import scala.scalajs.js.annotation.{JSExport, JSExportAll}
+import scala.scalajs.js.annotation.{JSExportTopLevel, JSExportAll}
 
-@JSExport @JSExportAll
+@JSExportTopLevel(name="sjs.example.Library") @JSExportAll
 object Library {
   def foo(): String = SomeOtherCode.quux(true)
   def bar(): String = SomeOtherCode.quux(false)