+++ sbt-scalajs-bundler/src/sbt-test/sbt-scalajs-bundler/library/webpack.config.js
@@ -1,3 +1,0 @@
-module.exports = require('./scalajs.webpack.config');
-// Expose the Scala.js artifact as a library in the `sjs` nampespace
-module.exports.output.library = 'sjs';