+++ sbt-web-scalajs-bundler/src/main/scala/scalajsbundler/sbtplugin/WebScalaJSBundlerPlugin.scala
@@ -9,6 +9,7 @@ import webscalajs.WebScalaJS
 import webscalajs.WebScalaJS.autoImport._
 
 import scala.collection.immutable.Nil
+import scalajsbundler.Webpack
 import scalajsbundler.sbtplugin.ScalaJSBundlerPlugin.autoImport._
 
 /**
@@ -17,7 +18,7 @@ import scalajsbundler.sbtplugin.ScalaJSBundlerPlugin.autoImport._
   *
   * = Tasks and Settings =
   *
-  * See the [[WebScalaJSBundlerPlugin$.autoImport autoImport]] member.
+  * See the [[WebScalaJSBundlerPlugin.autoImport autoImport]] member.
   */
 object WebScalaJSBundlerPlugin extends AutoPlugin {
 
@@ -29,7 +30,7 @@ object WebScalaJSBundlerPlugin extends AutoPlugin {
     /**
       * Sequence of PathMapping’s to include to sbt-web’s assets.
       *
-      * @see [[scalajsbundler.sbtplugin.NpmAssets$.ofProject NpmAssets.ofProject]]
+      * @see [[scalajsbundler.sbtplugin.NpmAssets.ofProject NpmAssets.ofProject]]
       * @group settings
       */
     val npmAssets: TaskKey[Seq[PathMapping]] = taskKey[Seq[PathMapping]]("Assets (resources that are not CommonJS modules) imported from the NPM packages")
@@ -60,13 +61,14 @@ object WebScalaJSBundlerPlugin extends AutoPlugin {
     )
   }
 
-
   def pipelineStage(sjsStage: TaskKey[Attributed[File]], self: TaskKey[Pipeline.Stage]): Def.Initialize[Task[Pipeline.Stage]] =
     Def.taskDyn {
       val projects = scalaJSProjects.value.map(Project.projectToRef)
       val npmAssetsMappings = npmAssets.value
+      val selfIncludeFilter = (includeFilter in self).value
+      val selfExcludeFilter = (excludeFilter in self).value
       Def.task { mappings: Seq[PathMapping] =>
-        val filtered = filterMappings(mappings, (includeFilter in self).value, (excludeFilter in self).value)
+        val filtered = filterMappings(mappings, selfIncludeFilter, selfExcludeFilter)
 
         // ((file, relative-path), sourceMapsEnabled)
         val bundles: Seq[((File, String), Boolean)] =
@@ -76,16 +78,16 @@ object WebScalaJSBundlerPlugin extends AutoPlugin {
                 val files = (webpack in (project, Compile, sjsStage in project)).value
                 val clientTarget = (npmUpdate in (project, Compile)).value
                 val sourceMapsEnabled = (webpackEmitSourceMaps in (project, Compile, sjsStage in project)).value
-                files.pair(relativeTo(clientTarget)).map((_, sourceMapsEnabled))
+                files.map(_.data).pair(relativeTo(clientTarget)).map((_, sourceMapsEnabled))
               }
             }
             .foldLeft(Def.task(Seq.empty[((File, String), Boolean)]))((acc, bundleFiles) => Def.task(acc.value ++ bundleFiles.value))
             .value
 
         val bundlesWithSourceMaps =
-          bundles.flatMap { case ((file, path), sourceMapsEnabled) =>
-            if (sourceMapsEnabled) {
-              val sourceMapFile = file.getParentFile / (file.name ++ ".map")
+          bundles.flatMap { case ((file, path), bundleSourceMapsEnabled) =>
+            val sourceMapFile = file.getParentFile / (file.name ++ ".map")
+            if (bundleSourceMapsEnabled && sourceMapFile.exists) {
               val sourceMapPath = path ++ ".map"
               Seq(file -> path, sourceMapFile -> sourceMapPath)
             } else Seq(file -> path)